Memory usage will now optionally be displayed on the top of the screen.
The label will be shown at the top center once the capturing process is initiated, signalling the recording is on and the MB of RAM being used by the resource. The saved JSON file will generally be half of that value.
This label can be hidden by the command /memory.
The refresh rate of the memory count can be set with /memrefresh.

Also:
- Updated driver ped capture to also include the root position and rotation ...
- ... and bone rotations are now saved as mostly Quaternion, except 3 digit ID bones that are still saved as Euler.